Update product route
