+++ app/utils/product.ts
@@ -1,4 +1,9 @@
+import type { Storefront } from "@shopify/hydrogen";
 import type { MoneyV2 } from "@shopify/hydrogen/storefront-api-types";
+import type { ProductRecommendationsQuery } from "storefrontapi.generated";
+import invariant from "tiny-invariant";
+import { RECOMMENDED_PRODUCTS_QUERY } from "~/graphql/queries";
+import type { I18nLocale } from "~/types/locale";
 
 export function isNewArrival(date: string, daysOld = 30) {
   return (
@@ -12,4 +17,32 @@ export function isDiscounted(price: MoneyV2, compareAtPrice: MoneyV2) {
     return true;
   }
   return false;
+}
+
+export async function getRecommendedProducts(
+  storefront: Storefront<I18nLocale>,
+  productId: string,
+) {
+  let products = await storefront.query<ProductRecommendationsQuery>(
+    RECOMMENDED_PRODUCTS_QUERY,
+    {
+      variables: { productId, count: 12 },
+    },
+  );
+
+  invariant(products, "No data returned from Shopify API");
+
+  let mergedProducts = (products.recommended ?? [])
+    .concat(products.additional.nodes)
+    .filter((prod, idx, arr) => {
+      return arr.findIndex(({ id }) => id === prod.id) === idx;
+    });
+
+  let originalProduct = mergedProducts.findIndex(
+    (item) => item.id === productId,
+  );
+
+  mergedProducts.splice(originalProduct, 1);
+
+  return { nodes: mergedProducts };
 }
